% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_assignment_class.R
\name{assignmentClass}
\alias{assignmentClass}
\title{Assignment Classes & Methods}
\usage{
assignmentClass(x)
}
\arguments{
\item{x}{list returned by api call}
}
\description{
Assignments class extends the data.table class to include relevant methods
}
