% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveySolutionsAPI.R
\docType{package}
\name{SurveySolutionsAPI}
\alias{SurveySolutionsAPI}
\alias{SurveySolutionsAPI-package}
\title{Survey Solutions API package}
\description{
This is a first version of an R package, to access the Survey Solutions RESTful API,
it allows you to export data/paradata, manipulate users and questionnaires, as well as
to create and manipulate assignments directly out of R.
}
\details{
The package is not on CRAN yet, so you have to install it by using.

If you want to contribute or have suggestions, please send an email to the Survey Solutions
team
}
