% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_user.R
\name{suso_archUSR}
\alias{suso_archUSR}
\title{Survey Solutions API call archive user}
\usage{
suso_archUSR(
  url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"),
  uid = NULL,
  archive = F,
  workspace = NULL,
  token = NULL
)
}
\arguments{
\item{url}{Survey Solutions server address}

\item{usr}{Survey Solutions API user}

\item{pass}{Survey Solutions API password}

\item{uid}{user id}

\item{archive}{if TRUE user will be archived or statys archived, if FALSE user will be unarchived or stays unarchived}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace, if
no workspace is provided defaults to primary workspace.}

\item{token}{If Survey Solutions server token is provided \emph{usr} and \emph{pass} will be ignored}
}
\description{
(Un-)Archive user
}
