% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_workspace.R
\name{suso_assignWorkspace}
\alias{suso_assignWorkspace}
\title{Survey Solutions API call to assign workspace}
\usage{
suso_assignWorkspace(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  token = NULL,
  assign_workspace = NULL,
  your_workspace = NULL,
  uid = NULL,
  sv_id = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions ADMIN user}

\item{apiPass}{Survey Solutions ADMIN password}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{assign_workspace}{The workspace which you want to assign to the new user}

\item{your_workspace}{The workspace the API user has currently access to.}

\item{uid}{The User ID of the user to be assigned.}

\item{sv_id}{The supervisor's ID to which the interviewer should be assigned to, if it is a supervisor who is assigned, just use the same
as in \emph{uid}.}
}
\value{
If succesfull, returns a data.table with the details as well as the Status message "Worspaces list updated".
}
\description{
\code{suso_assignWorkspace} Allows you to assign a workspace to a specific user.
For more details please read \url{https://docs.mysurvey.solutions/headquarters/accounts/workspaces/}. To run this command
you require admin credentials.
}
\details{
Be aware, that for using this call you require admin credentials, and not the regular API user credentials.
}
\examples{
\dontrun{
# Use Admin Credentials!
suso_assignWorkspace(
          your_workspace = "myworkspace",
          assign_workspace = "myworkspace1",
          uid = "xxx-xxx-xxx-xxx-xxx",
          sv_id = "xxx-xxx-xxx-xxx-xxx",
          apiUser = "xxxxxx",
          apiPass = "xxxxxx")
}


}
