% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_create.R
\name{suso_createASS}
\alias{suso_createASS}
\title{Survey Solutions API call for Assignment Creation}
\usage{
suso_createASS(df = UPLOADdataCHECK,
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), QUID = NULL,
  version = NULL)
}
\arguments{
\item{df}{dataframe with upload data}

\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{QUID}{the questionnaire id}

\item{version}{the questionnaire version}
}
\description{
Creates assignment with prefilled data.
}
\details{
Dataframe needs to be provided with columns
for prefilling data, as well as \emph{Quantity} and \emph{ResponsibleName}. Return value is a data.table,
with the prefilling data, if successful
}
