% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_create_user.R
\name{suso_createUSER}
\alias{suso_createUSER}
\title{Survey Solutions API call for User Creation}
\usage{
suso_createUSER(
  userlist = NULL,
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = NULL,
  token = NULL,
  showUser = T
)
}
\arguments{
\item{userlist}{dataframe with upload data}

\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{showUser}{userName in console}
}
\description{
Creates Survey Solutions users (interviewers or supervisors).
}
\details{
Dataframe needs to be provided with the mandatory columns
for user creation which are: Role, UserName, Password and Supervisor (in case of interviewer),
optional you can also provide FullName, PhoneNumber and Email. Return value is a data.table, which includes
the user information as well as the response's status code. Important is also that the UserName and Password are
provided in the required format.
}
