% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_export_details.R
\name{suso_details_lastexport}
\alias{suso_details_lastexport}
\title{Helper function for details (i.e. date, compilation status etc. )}
\usage{
suso_details_lastexport(url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"), quid = "", version = 1,
  format = "STATA")
}
\arguments{
\item{url}{Survey Solutions server address}

\item{usr}{Survey Solutions API user}

\item{pass}{Survey Solutions API password}

\item{quid}{Questionnaire ID}

\item{version}{Questionnaire version}

\item{format}{one of Tab, Csv or Excel}
}
\description{
This function retrieves the details from last file creation and can be used to modify
download file creation behavior.
}
