% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_export_files.R
\name{suso_export}
\alias{suso_export}
\title{Survey Solutions API call to generate and download the data}
\usage{
suso_export(questName = "PAPEL - ENCOVI Cuestionario de Listado FINAL",
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  questID = "e4de521a-6e32-4ab0-93c5-1fa4e11dc12f", version = 2,
  workStatus = "Completed", reloadTimeDiff = 1, inShinyApp = F,
  n_id_vars = 11)
}
\description{
It has the following characteristics.
 \itemize{
   \item Function returns a LIST with up to 4 different lists
   \item Number of lists depends on the level of roster nesting
   \item All variable names are transformed to lower case and categorical variables are consistently labeled
   \item Consistent id variables are generated with
         \itemize{
         \item interview__id transformed to id
         \item parent ids consistently number starting from id (questionnairid) to idX (maximum id3)
         }
   \item List elements are returnde as data.tables
   \item Allows for specification of reload time (i.e. generation of new download file)
   }
}
\details{
@param workStatus define which statuses the file should inlude, if NULL all is exported
   @param reloadTimeDiff time difference in hours between last generated file and now
   @param inShinyApp if True, file interacts with shiny progress bar
   @param n_id_vars  specify the number of identification variables

@export
}
