% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_credentials.R
\name{suso_get_api_key}
\alias{suso_get_api_key}
\title{Get credentials}
\usage{
suso_get_api_key(api = c("susoServer", "susoUser", "susoPass"))
}
\arguments{
\item{api}{One of susoServer, susoUser, susoPass}
}
\description{
Get API credentials
}
\details{
Get credentials, used as input in API calls
}
