% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_assignment.R
\name{suso_get_assignments}
\alias{suso_get_assignments}
\title{Survey Solutions API call for assignment list}
\usage{
suso_get_assignments(questID = "",
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), AssId = NULL,
  version = NULL, responsibleID = NULL,
  order.by = c("Id,ResponsibleId,ResponsibleName,QuestionnaireId,InterviewsCount,\\n                                          Quantity,Archived,CreatedAtUtc,UpdatedAtUtc,\\n                                          IdentifyingQuestions,Email,Password,WebMode"),
  supervisorId = NULL, quantity = NULL)
}
\arguments{
\item{questID}{only assignments for questionnaire ID are returned, requires \code{version} being not NULL}

\item{AssId}{if not NULL a list of all assignments on the server, if NULL
the assignment details for a specific assignment ID}

\item{version}{version of the questionnaire, only required with \code{questID}}

\item{responsibleID}{the ID of the responsible user (Supervisor or Interviewer).
Retrieves all assignments for this user.}

\item{order.by}{determines the column by which the assignment list should be ordered}
}
\description{
\code{suso_get_assignments} calls the Survey Solutions assingment API
}
