% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_credentials.R
\name{suso_get_default_key}
\alias{suso_get_default_key}
\title{Checks if credentials are present}
\usage{
suso_get_default_key(api = c("susoServer", "susoUser", "susoPass"))
}
\arguments{
\item{api}{one of susoServer, susoUser or susoPass}
}
\description{
Helper function
}
