% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_get_stats_interview}
\alias{suso_get_stats_interview}
\title{Get statistics for interview}
\usage{
suso_get_stats_interview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = NULL,
  token = NULL,
  intID = ""
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{intID}{a single or multiple \emph{InterviewId}.}
}
\description{
Get statistics for interview
}
