% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_getAllAnswerInterview}
\alias{suso_getAllAnswerInterview}
\title{Get all answers for a specific interview}
\usage{
suso_getAllAnswerInterview(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), intID = "")
}
\description{
Get all answers for a specific interview
}
