% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_getAllHistoryInterview}
\alias{suso_getAllHistoryInterview}
\title{Get all history for a specific interview}
\usage{
suso_getAllHistoryInterview(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), intID = "")
}
\description{
Get all history for a specific interview
}
