% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_user.R
\name{suso_getINT}
\alias{suso_getINT}
\title{Survey Solutions API call for list of interviewers}
\usage{
suso_getINT(url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"), sv_id = NULL)
}
\description{
Survey Solutions API call for list of interviewers
}
