% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_statistics.R
\name{suso_getStatsQuestionnaire}
\alias{suso_getStatsQuestionnaire}
\title{Survey Solutions API call for variable specific statistics
ATTENTION: Test Version, will be updated soon}
\usage{
suso_getStatsQuestionnaire(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), questID = "", version = "",
  qQuest = "", qExp = "Csv")
}
\description{
Survey Solutions API call for variable specific statistics
ATTENTION: Test Version, will be updated soon
}
