% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_user.R
\name{suso_getSV}
\alias{suso_getSV}
\title{Survey Solutions API call for list of supervisors}
\usage{
suso_getSV(url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"))
}
\arguments{
\item{url}{Survey Solutions server address}

\item{usr}{Survey Solutions API user}

\item{pass}{Survey Solutions API password}
}
\description{
Gets list of supervisors
}
