% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_credentials.R
\name{suso_keys}
\alias{suso_keys}
\title{Survey Solutions API credentials}
\usage{
suso_keys()
}
\description{
(this function has been inspired by the googleway package
\url{https://github.com/SymbolixAU/googleway},
an excellent package to use google geo-spatial API)
Retrieves the list of Survey Solutions credentials that have been set.
}
