% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_maps.R
\name{suso_mapinfo}
\alias{suso_mapinfo}
\title{Receive maps currently uploaded to the server}
\usage{
suso_mapinfo(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = NULL,
  fileName = NULL,
  importDateUtc = NULL,
  size = NULL,
  users = NULL,
  sortby_filename = NULL,
  sortby_importeddateutc = NULL,
  sortby_size = NULL,
  take = NULL,
  skip = NULL
)
}
\arguments{
\item{server}{GraphQL endpoint of your server}

\item{apiUser}{your API username}

\item{apiPass}{API password}

\item{workspace}{Server Workspace, if NULL uses default}

\item{fileName}{name of the map on the server}

\item{importDateUtc}{Import date}

\item{size}{Size of the map}

\item{users}{Users to whom the maps are assigned}

\item{sortby_filename}{sort maps by file name, either ASC for ascending or DESC for descending}

\item{sortby_importeddateutc}{sort maps by import date in utc, either ASC for ascending or DESC for descending}

\item{sortby_size}{sort by map size, either ASC for ascending or DESC for descending}

\item{take}{take the specified integer numeber of maps}

\item{skip}{skip the first integer number of maps}
}
\value{
Returns a data.table, with all the maps and additonal information. If multiple users are assigned to a map,
the table is expanded, such that there is one user per map.
}
\description{
Allows the user to retrieve filtered or unfiltered map data.
}
\details{
Attention: this uses the GraphQL API, not the REST API.
}
\examples{
\dontrun{
suso_mapinfo(workspace = "myworkspace")
}

}
