% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_maps.R
\name{suso_mapupload}
\alias{suso_mapupload}
\title{Upload map to server}
\usage{
suso_mapupload(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = NULL,
  token = NULL,
  path_to_zip = NULL
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{path_to_zip}{either path to zip file or directory of files which will automatically be zipped as required}
}
\value{
Returns a data.table, with information about the uploaded maps
}
\description{
Allows the user to upload a zip file of background maps and boundary files
}
\details{
Resources for upload must meet the requirements specified under:
\url{https://docs.mysurvey.solutions/headquarters/mapsmanage/map-formats/},
Attention: this uses the GraphQL API, not the REST API.
}
