% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_patchRejectInterview}
\alias{suso_patchRejectInterview}
\title{Reject interview}
\usage{
suso_patchRejectInterview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  intID = "",
  comment = "Please check errors and re-submit!"
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{intID}{the \emph{InterviewId} of the interview.}

\item{comment}{comment which should be sent with the questionnaire}
}
\description{
N.B. under development
}
