% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_credentials.R
\name{suso_PwCheck}
\alias{suso_PwCheck}
\title{Utility function to check if credentials are correct}
\usage{
suso_PwCheck(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"))
}
\arguments{
\item{server}{Survey Solutions Server}

\item{apiUser}{API user}

\item{apiPass}{API password}
}
\description{
This function returns a 200 status if the correct credentials have been provided.
}
