% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_assignment.R
\name{suso_set_assignments}
\alias{suso_set_assignments}
\title{Survey Solutions API call for assingment manipulation}
\usage{
suso_set_assignments(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), AssId = NULL,
  responsibleID = NULL)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{AssId}{if NULL a list of all assignments on the server, if not NULL
the assignment details for a specific assignment ID}

\item{responsibleID}{the ID of the responsible user (Supervisor or Interviewer).
Retrieves all assignments for this user.}
}
\description{
\code{suso_set_assignments} allows to (re-)assign, change limits or audio
recording settings, as well as archiving assignments.
ATTENTION: still under development.
}
