% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_credentials.R
\name{suso_set_key}
\alias{suso_set_key}
\title{Set Credentials}
\usage{
suso_set_key(suso_server = "", suso_user = "", suso_password = "")
}
\arguments{
\item{suso_server}{Survey Solutions server address}

\item{suso_user}{Survey Solutions API user}

\item{suso_password}{Survey Solutions API password}
}
\description{
Sets API credentials so it's available for all API calls. See details
}
\details{
Use \code{suso_set_key} to make API keys available for all the \code{suso_}
functions, so you don't need to specify the credentials parameter within those
functions.
}
