% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Questionnaire.R
\name{suso_transform_fullMeta}
\alias{suso_transform_fullMeta}
\title{Transforms list of questionnair structure to single data.table with all variables}
\usage{
suso_transform_fullMeta(input = NULL)
}
\arguments{
\item{input}{returned by \code{suso_getQuestDetails} structure operation}
}
\description{
\code{suso_transform_fullMeta} transforms the list containing the structure (\emph{operation.type = structure})
is transformed into a single data.table
with all variable names, types etc.. This also works with json strings manually exported from the server.
}
