% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_create.R
\name{suso_createASS}
\alias{suso_createASS}
\title{Survey Solutions API call for Assignment Creation}
\usage{
suso_createASS(df = UPLOADdataCHECK,
  url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"), QUID = NULL, version = NULL)
}
\description{
ATTENTION: NOT CHECKED YET
}
