% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_export_details.R
\name{suso_details_lastexport}
\alias{suso_details_lastexport}
\title{Helper function for details (i.e. date, compilation status etc. )}
\usage{
suso_details_lastexport(url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"), quid = "", version = 1,
  format = "STATA", q_name = "questionnaire name")
}
\description{
This function retrieves the details from last file creation and can be used to modify
download file creation behavior.
}
