% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_export_files.R
\name{suso_export}
\alias{suso_export}
\title{Survey Solutions API call to generate and download the data}
\usage{
suso_export(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = NULL,
  token = NULL,
  questID = "",
  version = 1,
  workStatus = "Completed",
  reloadTimeDiff = 1,
  inShinyApp = F,
  n_id_vars = 11
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{questID}{Questionnaire ID}

\item{version}{Questionnaire version}

\item{workStatus}{define which statuses the file should inlude (i.e. \emph{Restored,Created,SupervisorAssigned,InterviewerAssigned,
RejectedBySupervisor,ReadyForInterview,
SentToCapi,Restarted,Completed,ApprovedBySupervisor,
RejectedByHeadquarters,ApprovedByHeadquarters,Deleted}), if NULL all is exported}

\item{reloadTimeDiff}{time difference in hours between last generated file and now}

\item{inShinyApp}{if True, file interacts with shiny progress bar}

\item{n_id_vars}{specify the number of identification variables (not used for now!)}
}
\description{
Generates and downloads the data from your Survey Solutions server.
}
\details{
This API call uses the STATA export format to retrieve the categorical variables and labels.The result
has the following characteristics:
\itemize{
  \item it is returned as a LIST with up to 4 different lists. The list names are:
          \itemize{
             \item \emph{main} Contains the top level data, and (if available interviewer comments)
             \item \emph{R1} All rosters in roster level 1
             \item \emph{R2} All rosters in roster level 2
             \item \emph{R3} All rosters in roster level 3
          }
  \item Number of lists depends on the level of roster nesting
  \item All variable names are transformed to lower case and categorical variables are consistently labeled
  \item Consistent id variables are generated with
         \itemize{
         \item interview__id transformed to id
         \item parent ids consistently number starting from id (questionnairid) to idX (maximum id3)
         }
  \item List elements are returned as data.tables
  \item Allows for specification of reload time (i.e. generation of new download file)
  \item PRESERVES categorical labels \emph{and} values.
  }
}
