% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_paradata_v2.R
\name{suso_export_paradata}
\alias{suso_export_paradata}
\title{SURVEY SOLUTIONS PARADATA EXPORT FUNCTION}
\usage{
suso_export_paradata(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  questID = "xxxx-xxx-xxxx-xxx-xxx", version = 1, user = user(),
  inShinyServer = FALSE)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{questID}{\emph{QuestionnaireId} for which the paradata should be generated}

\item{version}{questionnnaire version}

\item{user}{only for specific user}

\item{inShinyServer}{if TRUE communicates with shiny server progress bar}
}
\description{
Exports Survey Solutions Paradata, and returns a data.table.
}
\details{
\code{suso_export_paradata} returns a data.table. Calculates the response time
 and separtes multiple responses into individual columns. It also creates a variable
 \emph{counter} which preserves the sequence of events.
}
