% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_getAllAnswerInterview}
\alias{suso_getAllAnswerInterview}
\title{Survey Solutions API call to retrieve all answers for a specific interview}
\usage{
suso_getAllAnswerInterview(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  intID = ""
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{intID}{the \emph{InterviewId} of the interview. To get a list of all interview for a specific questionnaire, execute \code{suso_getAllInterviewQuestionnaire}}
}
\description{
Returns all responses for a specific interview
}
