% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_getAllInterviewQuestionnaire}
\alias{suso_getAllInterviewQuestionnaire}
\title{Get all interviews for a specific questionnaire}
\usage{

  suso_getAllInterviewQuestionnaire(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  questID = "e4de521a-6e32-4ab0-93c5-1fa4e11dc12f", version = 2,
  status = NULL)
}
\description{
Get all interviews for a specific questionnaire
}
