% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_getAllInterviewQuestionnaire}
\alias{suso_getAllInterviewQuestionnaire}
\title{Survey Solutions API call to retrieve all interviews for a specific questionnaire}
\usage{
suso_getAllInterviewQuestionnaire(
  server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  workspace = NULL,
  token = NULL,
  questID = "",
  version = 1,
  workStatus = "Completed"
)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{workspace}{server workspace, if nothing provided, defaults to primary}

\item{token}{If Survey Solutions server token is provided \emph{apiUser} and \emph{apiPass} will be ignored}

\item{questID}{your Survey Solutions \emph{QuestionnaireId}. Retrieve a list of questionnaires by executing \code{suso_getQuestDetails}}

\item{version}{version of the questionnaire}

\item{workStatus}{define which statuses the file should inlude (i.e. \emph{Restored,Created,SupervisorAssigned,InterviewerAssigned,
RejectedBySupervisor,ReadyForInterview,
SentToCapi,Restarted,Completed,ApprovedBySupervisor,
RejectedByHeadquarters,ApprovedByHeadquarters,Deleted}), if NULL only completed interviews will be shown.}
}
\description{
Returns all interviews for the specified questionnaire and the selected status.
}
\details{
ATTENTION: This function will soon be switched to the GraphQL endpoint, since the REST API is
already deprecated.
}
\examples{
\dontrun{
To get all interviews for a specific questionnaire and a specific status
     suso_getAllInterviewQuestionnaire(
              workspace = "myworkspace",
              questID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea",
              version = 1,
              workStatus = "InterviewerAssigned"
              )
# or to get all interviews in status completed
     suso_getAllInterviewQuestionnaire(
              workspace = "myworkspace",
              questID = "dee7705f-d611-4b12-9b97-2b8e5b80c4ea",
              version = 1
              )

}


}
