% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Interviews.R
\name{suso_getAllInterviewQuestionnaire}
\alias{suso_getAllInterviewQuestionnaire}
\title{Survey Solutions API call to retrieve all interviews for a specific questionnaire}
\usage{

  suso_getAllInterviewQuestionnaire(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"),
  questID = "e4de521a-6e32-4ab0-93c5-1fa4e11dc12f", version = 2,
  status = NULL)
}
\arguments{
\item{questID}{your Survey Solutions \emph{QuestionnaireId}. Retrieve a list of questionnaires by executing \code{suso_getQuestDetails}}

\item{version}{version of the questionnaire}
}
\description{
Returns all interviews for the specific questionnaire.
}
