% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_user.R
\name{suso_getINT}
\alias{suso_getINT}
\title{Survey Solutions API call for list of interviewers}
\usage{
suso_getINT(
  url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"),
  sv_id = NULL,
  workspace = NULL,
  token = NULL
)
}
\arguments{
\item{url}{Survey Solutions server address}

\item{usr}{Survey Solutions API user}

\item{pass}{Survey Solutions API password}

\item{sv_id}{supervisor id}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace, if
no workspace is provided defaults to primary workspace.}

\item{token}{If Survey Solutions server token is provided \emph{usr} and \emph{pass} will be ignored}
}
\description{
Get list of all interviewers by supervisor
}
\examples{
\dontrun{
suso_getINT(
          workspace = "myworkspace",
          sv_id = "xxxx-xxxx-xxxx-xxx"
          )
}

}
