% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_user.R
\name{suso_getINT_log}
\alias{suso_getINT_log}
\title{Survey Solutions API call for audit log on interviewers}
\usage{
suso_getINT_log(url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"), int_id = NULL)
}
\arguments{
\item{url}{Survey Solutions server address}

\item{usr}{Survey Solutions API user}

\item{pass}{Survey Solutions API password}

\item{int_id}{interviewer id}
}
\description{
Get log of interviewer
}
