% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Questionnaire.R
\name{suso_getQuestDetails}
\alias{suso_getQuestDetails}
\title{Survey Solutions API call for questionnaire}
\usage{
suso_getQuestDetails(url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"), quid = NULL, version = NULL,
  limit = 40, offset = 0, operation.type = c("list", "statuses",
  "structure", "interviews"))
}
\description{
\code{suso_getQuestDetails} implements all Questionnaire API commands. If \emph{operation.type = structure}
the return file will be a list, whith the content of the raw json string containing the questionnaire variables
}
