% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_Questionnaire.R
\name{suso_getQuestDetails}
\alias{suso_getQuestDetails}
\title{Survey Solutions API call for questionnaire}
\usage{
suso_getQuestDetails(url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"), quid = NULL, version = NULL,
  operation.type = c("list", "statuses", "structure", "interviews"))
}
\arguments{
\item{url}{Survey Solutions server address}

\item{usr}{Survey Solutions API user}

\item{pass}{Survey Solutions API password}

\item{quid}{\emph{QuestionnaireId} for which details should be exported}

\item{version}{questionnaire version}

\item{operation.type}{if \emph{list} is specified a list of all questionnaires on the server. If
\emph{statuses} a vector of all questionnaire statuses. If \emph{structure} is specified, it returns a list
containing the content of the questionnaire's json string. To make use of it, put the output into
\code{suso_transform_fullMeta} and receive a data.table with all questionnaire elements. If \emph{interviews}, all interviews
for a specific questionnaire.}
}
\description{
\code{suso_getQuestDetails} implements all Questionnaire API commands. If \emph{operation.type = structure}
the return file will be a list, whith the content of the raw json string containing the questionnaire variables
}
