% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_statistics.R
\name{suso_getQuestionsQuestionnaire}
\alias{suso_getQuestionsQuestionnaire}
\title{Survey Solutions API call for questions and responses from single questionnaire}
\usage{
suso_getQuestionsQuestionnaire(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), questID = NULL,
  version = NULL)
}
\arguments{
\item{server}{Survey Solutions server address}

\item{apiUser}{Survey Solutions API user}

\item{apiPass}{Survey Solutions API password}

\item{questID}{only assignments for \emph{QuestionnaireId} are returned, requires \code{version} being not NULL}

\item{version}{version of the questionnaire, only required with \code{questID}}
}
\description{
Returns questions for single questionnaire (if they contain responses)
}
