% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_user.R
\name{suso_getUSR}
\alias{suso_getUSR}
\title{Survey Solutions API call any user}
\usage{
suso_getUSR(
  url = suso_get_api_key("susoServer"),
  usr = suso_get_api_key("susoUser"),
  pass = suso_get_api_key("susoPass"),
  uid = NULL,
  workspace = NULL,
  token = NULL
)
}
\arguments{
\item{url}{Survey Solutions server address}

\item{usr}{Survey Solutions API user}

\item{pass}{Survey Solutions API password}

\item{uid}{user id}

\item{workspace}{If workspace name is provide requests are made regarding this specific workspace, if
no workspace is provided defaults to primary workspace.}

\item{token}{If Survey Solutions server token is provided \emph{usr} and \emph{pass} will be ignored}
}
\description{
Get any user's info
}
\examples{
\dontrun{
suso_getUSR(
          workspace = "myworkspace",
          uid = "xxxx-xxxx-xxxx-xxx"
          )
}

}
