% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suso_api_assignment.R
\name{suso_set_assignments}
\alias{suso_set_assignments}
\title{Survey Solutions API call for assingment manipulation}
\usage{
suso_set_assignments(server = suso_get_api_key("susoServer"),
  apiUser = suso_get_api_key("susoUser"),
  apiPass = suso_get_api_key("susoPass"), AssId = NULL,
  responsibleID = NULL)
}
\description{
\code{suso_set_assignments} allows to (re-)assign, change limits or audio
recording settings, as well as archiving assignments.
ATTENTION: still under development.
}
